Test cleanup.
JobStore and Repository tests can now bootstrap and clean up all tables easily.
