--- src/QuartzNET-DynamoDB.Tests/Integration/TestJobStoreFactory.cs
@@ -1,16 +1,0 @@
-﻿using System.Configuration;
-namespace Quartz.DynamoDB.Tests
-{
-    public class TestJobStoreFactory
-    {
-        private static JobStore _store;
-
-        public static DynamoDB.JobStore CreateTestJobStore()
-        {
-            _store = new JobStore();
-            _store.InstanceName = ConfigurationManager.AppSettings["InstanceName"];
-
-            return _store;
-        }
-    }
-}
